backend/requirements.txt
Flask==3.0.0                   # Web framework
Flask-Cors==4.0.0              # Enable CORS for frontend-backend communication
pymongo==4.6.1                 # MongoDB driver for Python
dnspython==2.3.0               # Required for connecting to MongoDB Atlas
flask-bcrypt==1.0.1            # Password hashing with bcrypt
pyjwt==2.8.0                   # JWT (JSON Web Token) for authentication
python-dotenv
web3
ipfshttpclient
python-multipart